Split AssemblyLoadContext run test from Library tuple serialization test
